output/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.1
python-dotenv==1.0.0
colorama==0.4.6
